Revert "feat(core): re-introduce nested leave animations scoped to component boundaries"
This reverts commit df659b8d0cf64eeed418c60bc16cae5630086401.
