+++ integration/animations/src/app/app.component.ts
@@ -1,15 +1,45 @@
 import {Component} from '@angular/core';
-import {RouterLink, RouterOutlet} from '@angular/router';
+import {
+  CdkDragDrop,
+  CdkDropList,
+  CdkDrag,
+  CdkDragPlaceholder,
+  moveItemInArray,
+} from '@angular/cdk/drag-drop';
+
+// We want to verify that dragging an item does not result in any items disappearing
+// when they have an enter/leave animation.
 
 @Component({
   selector: 'app-root',
-  imports: [RouterOutlet, RouterLink],
-  template: `
-    <nav>
-      <a routerLink="/" id="home-link">Home</a> |
-      <a routerLink="/nested" id="nested-link">Nested Animations</a>
-    </nav>
-    <router-outlet></router-outlet>
-  `,
+  imports: [CdkDropList, CdkDrag, CdkDragPlaceholder],
+  templateUrl: './app.component.html',
+  styleUrl: './app.component.css',
 })
-export class AppComponent {}
+export class AppComponent {
+  movies = [
+    'Episode I - The Phantom Menace',
+    'Episode II - Attack of the Clones',
+    'Episode III - Revenge of the Sith',
+  ];
+
+  showFallback = true;
+
+  drop(event: CdkDragDrop<string[]>) {
+    moveItemInArray(this.movies, event.previousIndex, event.currentIndex);
+  }
+
+  hideAndIntercept() {
+    const el = document.querySelector('.fallback-el');
+    if (el) {
+      el.addEventListener(
+        'animationend',
+        (e) => {
+          e.stopImmediatePropagation();
+        },
+        true,
+      );
+    }
+    this.showFallback = false;
+  }
+}
