+++ integration/animations/src/main.ts
@@ -1,8 +1,4 @@
 import {bootstrapApplication} from '@angular/platform-browser';
 import {AppComponent} from './app/app.component';
-import {provideRouter} from '@angular/router';
-import {routes} from './app/app.routes';
 
-bootstrapApplication(AppComponent, {
-  providers: [provideRouter(routes)],
-}).catch((err) => console.error(err));
+bootstrapApplication(AppComponent).catch((err) => console.error(err));
