+++ packages/core/src/animation/queue.ts
@@ -24,14 +24,11 @@ export const ANIMATION_QUEUE = new InjectionToken<AnimationQueue>(
   typeof ngDevMode !== 'undefined' && ngDevMode ? 'AnimationQueue' : '',
   {
     factory: () => {
-      const injector = inject(EnvironmentInjector);
-      const queue = new Set<VoidFunction>();
-      injector.onDestroy(() => queue.clear());
       return {
-        queue,
+        queue: new Set(),
         isScheduled: false,
         scheduler: null,
-        injector, // should be the root injector
+        injector: inject(EnvironmentInjector), // should be the root injector
       };
     },
   },
@@ -59,20 +56,6 @@ export function addToAnimationQueue(
     animationData?.detachedLeaveAnimationFns?.push(animationFns);
   }
   animationQueue.scheduler && animationQueue.scheduler(injector);
-}
-
-export function removeAnimationsFromQueue(
-  injector: Injector,
-  animationFns: VoidFunction | VoidFunction[],
-) {
-  const animationQueue = injector.get(ANIMATION_QUEUE);
-  if (Array.isArray(animationFns)) {
-    for (const animateFn of animationFns) {
-      animationQueue.queue.delete(animateFn);
-    }
-  } else {
-    animationQueue.queue.delete(animationFns);
-  }
 }
 
 export function removeFromAnimationQueue(injector: Injector, animationData: AnimationLViewData) {
