+++ packages/core/test/bundling/animations-standalone/bundle.golden_symbols.json
@@ -323,7 +323,6 @@
       "addServerStyles",
       "addToAnimationQueue",
       "addToEndOfViewTree",
-      "aggregateDescendantAnimations",
       "allLeavingAnimations",
       "allocExpando",
       "allocLFrame",
@@ -371,10 +370,8 @@
       "cleanUpView",
       "cloakAndComputeStyles",
       "cloakElement",
-      "collectAllViewLeaveAnimations",
       "collectNativeNodes",
       "collectNativeNodesInLContainer",
-      "collectNestedViewAnimations",
       "computeStaticStyling",
       "computeStyle",
       "concatStringsWithSpace",
@@ -465,7 +462,6 @@
       "executeCheckHooks",
       "executeContentQueries",
       "executeInitAndCheckHooks",
-      "executeLeaveAnimations",
       "executeOnDestroys",
       "executeTemplate",
       "executeViewQueryFn",
@@ -764,7 +760,6 @@
       "relativePath",
       "rememberChangeHistoryAndInvokeOnChangesHook",
       "remove",
-      "removeAnimationsFromQueue",
       "removeClass",
       "removeElements",
       "removeFromArray",
