+++ packages/core/test/bundling/create_component/bundle.golden_symbols.json
@@ -252,7 +252,6 @@
       "addToArray",
       "addToEndOfViewTree",
       "addViewToDOM",
-      "aggregateDescendantAnimations",
       "allLeavingAnimations",
       "allocExpando",
       "allocLFrame",
@@ -287,10 +286,8 @@
       "captureError",
       "checkStable",
       "cleanUpView",
-      "collectAllViewLeaveAnimations",
       "collectNativeNodes",
       "collectNativeNodesInLContainer",
-      "collectNestedViewAnimations",
       "computeStaticStyling",
       "concatStringsWithSpace",
       "config",
@@ -372,7 +369,6 @@
       "executeCheckHooks",
       "executeContentQueries",
       "executeInitAndCheckHooks",
-      "executeLeaveAnimations",
       "executeListenerWithErrorHandling",
       "executeOnDestroys",
       "executeTemplate",
@@ -625,7 +621,6 @@
       "relativePath",
       "rememberChangeHistoryAndInvokeOnChangesHook",
       "remove",
-      "removeAnimationsFromQueue",
       "removeElements",
       "removeFromArray",
       "removeLViewOnDestroy",
