+++ packages/core/test/bundling/defer/bundle.golden_symbols.json
@@ -296,7 +296,6 @@
       "addToArray",
       "addToEndOfViewTree",
       "addViewToDOM",
-      "aggregateDescendantAnimations",
       "allLeavingAnimations",
       "allocExpando",
       "allocLFrame",
@@ -332,10 +331,8 @@
       "checkStable",
       "classIndexOf",
       "cleanUpView",
-      "collectAllViewLeaveAnimations",
       "collectNativeNodes",
       "collectNativeNodesInLContainer",
-      "collectNestedViewAnimations",
       "computeStaticStyling",
       "concatStringsWithSpace",
       "config",
@@ -414,7 +411,6 @@
       "executeCheckHooks",
       "executeContentQueries",
       "executeInitAndCheckHooks",
-      "executeLeaveAnimations",
       "executeOnDestroys",
       "executeTemplate",
       "executeViewQueryFn",
@@ -664,7 +660,6 @@
       "registerPreOrderHooks",
       "rememberChangeHistoryAndInvokeOnChangesHook",
       "remove",
-      "removeAnimationsFromQueue",
       "removeFromArray",
       "removeLViewFromLContainer",
       "removeLViewOnDestroy",
