+++ packages/core/test/bundling/forms_reactive/bundle.golden_symbols.json
@@ -368,7 +368,6 @@
       "addToEndOfViewTree",
       "addValidators",
       "addViewToDOM",
-      "aggregateDescendantAnimations",
       "allLeavingAnimations",
       "allocExpando",
       "allocLFrame",
@@ -419,10 +418,8 @@
       "cleanUpView",
       "coerceToAsyncValidator",
       "coerceToValidator",
-      "collectAllViewLeaveAnimations",
       "collectNativeNodes",
       "collectNativeNodesInLContainer",
-      "collectNestedViewAnimations",
       "collectResidual",
       "collectStylingFromDirectives",
       "collectStylingFromTAttrs",
@@ -536,7 +533,6 @@
       "executeCheckHooks",
       "executeContentQueries",
       "executeInitAndCheckHooks",
-      "executeLeaveAnimations",
       "executeListenerWithErrorHandling",
       "executeOnDestroys",
       "executeSchedule",
@@ -916,7 +912,6 @@
       "relativePath",
       "rememberChangeHistoryAndInvokeOnChangesHook",
       "remove",
-      "removeAnimationsFromQueue",
       "removeElements",
       "removeFromArray",
       "removeLViewOnDestroy",
