+++ packages/core/test/bundling/forms_template_driven/bundle.golden_symbols.json
@@ -369,7 +369,6 @@
       "addToEndOfViewTree",
       "addValidators",
       "addViewToDOM",
-      "aggregateDescendantAnimations",
       "allLeavingAnimations",
       "allocExpando",
       "allocLFrame",
@@ -417,10 +416,8 @@
       "cleanUpView",
       "coerceToAsyncValidator",
       "coerceToValidator",
-      "collectAllViewLeaveAnimations",
       "collectNativeNodes",
       "collectNativeNodesInLContainer",
-      "collectNestedViewAnimations",
       "collectResidual",
       "collectStylingFromDirectives",
       "collectStylingFromTAttrs",
@@ -533,7 +530,6 @@
       "executeCheckHooks",
       "executeContentQueries",
       "executeInitAndCheckHooks",
-      "executeLeaveAnimations",
       "executeListenerWithErrorHandling",
       "executeOnDestroys",
       "executeSchedule",
@@ -913,7 +909,6 @@
       "relativePath",
       "rememberChangeHistoryAndInvokeOnChangesHook",
       "remove",
-      "removeAnimationsFromQueue",
       "removeElements",
       "removeFromArray",
       "removeLViewOnDestroy",
