+++ packages/core/test/bundling/hydration/bundle.golden_symbols.json
@@ -292,7 +292,6 @@
       "addServerStyles",
       "addToAnimationQueue",
       "addToEndOfViewTree",
-      "aggregateDescendantAnimations",
       "allLeavingAnimations",
       "allocExpando",
       "allocLFrame",
@@ -336,10 +335,8 @@
       "cleanupLView",
       "cleanupMatchingDehydratedViews",
       "clearElementContents",
-      "collectAllViewLeaveAnimations",
       "collectNativeNodes",
       "collectNativeNodesInLContainer",
-      "collectNestedViewAnimations",
       "computeStaticStyling",
       "concatStringsWithSpace",
       "config",
@@ -429,7 +426,6 @@
       "executeCheckHooks",
       "executeContentQueries",
       "executeInitAndCheckHooks",
-      "executeLeaveAnimations",
       "executeOnDestroys",
       "executeSchedule",
       "executeTemplate",
@@ -706,7 +702,6 @@
       "relativePath",
       "rememberChangeHistoryAndInvokeOnChangesHook",
       "remove",
-      "removeAnimationsFromQueue",
       "removeDehydratedView",
       "removeDehydratedViews",
       "removeElements",
