+++ packages/core/test/bundling/router/bundle.golden_symbols.json
@@ -407,7 +407,6 @@
       "addViewToDOM",
       "advanceActivatedRoute",
       "afterNextNavigation",
-      "aggregateDescendantAnimations",
       "allLeavingAnimations",
       "allocExpando",
       "allocLFrame",
@@ -451,10 +450,8 @@
       "checkStable",
       "classIndexOf",
       "cleanUpView",
-      "collectAllViewLeaveAnimations",
       "collectNativeNodes",
       "collectNativeNodesInLContainer",
-      "collectNestedViewAnimations",
       "collectQueryResults",
       "combineLatest",
       "combineLatestInit",
@@ -604,7 +601,6 @@
       "executeCheckHooks",
       "executeContentQueries",
       "executeInitAndCheckHooks",
-      "executeLeaveAnimations",
       "executeListenerWithErrorHandling",
       "executeOnDestroys",
       "executeSchedule",
@@ -1010,7 +1006,6 @@
       "relativePath",
       "rememberChangeHistoryAndInvokeOnChangesHook",
       "remove",
-      "removeAnimationsFromQueue",
       "removeElements",
       "removeFromArray",
       "removeLViewOnDestroy",
