+++ packages/core/test/bundling/standalone_bootstrap/bundle.golden_symbols.json
@@ -233,7 +233,6 @@
       "addServerStyles",
       "addToAnimationQueue",
       "addToEndOfViewTree",
-      "aggregateDescendantAnimations",
       "allLeavingAnimations",
       "allocExpando",
       "allocLFrame",
@@ -267,10 +266,8 @@
       "captureError",
       "checkStable",
       "cleanUpView",
-      "collectAllViewLeaveAnimations",
       "collectNativeNodes",
       "collectNativeNodesInLContainer",
-      "collectNestedViewAnimations",
       "computeStaticStyling",
       "concatStringsWithSpace",
       "config",
@@ -344,7 +341,6 @@
       "executeCheckHooks",
       "executeContentQueries",
       "executeInitAndCheckHooks",
-      "executeLeaveAnimations",
       "executeOnDestroys",
       "executeTemplate",
       "executeViewQueryFn",
@@ -559,7 +555,6 @@
       "relativePath",
       "rememberChangeHistoryAndInvokeOnChangesHook",
       "remove",
-      "removeAnimationsFromQueue",
       "removeElements",
       "removeFromArray",
       "removeLViewOnDestroy",
